Replace experimental patch with upstream fix
AP+11s-mesh was fixed upstream, backport that commit as well as the
preceding commit which allows to disable 802.11b legacy rates.
The reason for the latter patch to be backported as well is that
apart from being very useful in 2.4GHz environments it allows the
second patch which fixes AP+11s-mesh to apply cleanly without changes.

Signed-off-by: Daniel Golle <daniel@makrotopia.org>