fix repositories.conf
We use our own (more up-to-date) build of the routing feed, disable
it in the original repositories.conf.
As we are using the upstream release in terms of core, base and
packages, do not add our locally built duplicates.

Alltogether this change makes sure that packages occur only once in
the list and hence that list is useful for the GSoC auto-updater.
Thanks to Paul for reporting this!